+++ Areas/Admin/Controllers/CustomersController.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data;
 using System.Data.Entity;
 using System.Linq;
 using System.Net;
-using System.Web;
 using System.Web.Mvc;
 using HoangLongTH.Models;
 
@@ -12,30 +10,38 @@ namespace HoangLongTH.Areas.Admin.Controllers
 {
     public class CustomersController : Controller
     {
-        private MyStroreEntities db = new MyStroreEntities();
+        private MyStroreEntities1 db = new MyStroreEntities1();
 
         // GET: Admin/Customers
         public ActionResult Index()
         {
-            var customer = db.Customer.Include(c => c.User);
-            return View(customer.ToList());
+            var customers = db.Customer.Include(c => c.User);
+            return View(customers.ToList());
         }
 
         // GET: Admin/Customers/Details/5
         public ActionResult Details(int? id)
         {
             if (id == null)
-            {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
+
             Customer customer = db.Customer.Find(id);
             if (customer == null)
-            {
                 return HttpNotFound();
-            }
+
+            // Lấy danh sách đơn hàng của khách
+            ViewBag.Orders = db.Order
+                               .Include(o => o.OrderDetail)
+                               .Where(o => o.CustomerID == id)
+                               .ToList();
+
             return View(customer);
         }
 
+        // Không cho phép thêm, sửa, xóa khách hàng
+        // => Comment hoặc bỏ hẳn các hành động Create, Edit, Delete
+
+        /*
         // GET: Admin/Customers/Create
         public ActionResult Create()
         {
@@ -43,9 +49,6 @@ namespace HoangLongTH.Areas.Admin.Controllers
             return View();
         }
 
-        // POST: Admin/Customers/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "CustomerID,CustomerName,CustomerPhone,CustomerEmail,CustomerAddress,Username")] Customer customer)
@@ -65,21 +68,16 @@ namespace HoangLongTH.Areas.Admin.Controllers
         public ActionResult Edit(int? id)
         {
             if (id == null)
-            {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
+
             Customer customer = db.Customer.Find(id);
             if (customer == null)
-            {
                 return HttpNotFound();
-            }
+
             ViewBag.Username = new SelectList(db.User, "Username", "Password", customer.Username);
             return View(customer);
         }
 
-        // POST: Admin/Customers/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "CustomerID,CustomerName,CustomerPhone,CustomerEmail,CustomerAddress,Username")] Customer customer)
@@ -98,27 +96,28 @@ namespace HoangLongTH.Areas.Admin.Controllers
         public ActionResult Delete(int? id)
         {
             if (id == null)
-            {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
+
             Customer customer = db.Customer.Find(id);
             if (customer == null)
-            {
                 return HttpNotFound();
-            }
+
             return View(customer);
         }
 
-        // POST: Admin/Customers/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Customer customer = db.Customer.Find(id);
+            var customer = db.Customer.Find(id);
+            if (customer == null)
+                return HttpNotFound();
+
             db.Customer.Remove(customer);
             db.SaveChanges();
             return RedirectToAction("Index");
         }
+        */
 
         protected override void Dispose(bool disposing)
         {
